+++ demo/meeting/src/views/DeviceSetup/index.tsx
@@ -2,9 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React from 'react';
-import { Heading } from 'amazon-chime-sdk-component-library-react';
+import {
+  Heading,
+  DeviceSelection,
+} from 'amazon-chime-sdk-component-library-react';
 
-import DeviceSelection from '../../containers/DeviceSelection';
 import JoinMeetingDetails from '../../containers/MeetingJoinDetails';
 
 import { StyledLayout } from './Styled';