+++ src/components/sdk/MeetingControlsContainer/index.tsx
@@ -1,0 +1,14 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import AudioInputControl from './AudioInputControl';
+import VideoInputControl from './VideoInputControl';
+import ContentShareControl from './ContentShareControl';
+import AudioOutputControl from './AudioOutputControl';
+
+export { 
+  AudioInputControl, 
+  AudioOutputControl, 
+  VideoInputControl, 
+  ContentShareControl 
+};