+++ src/constants/index.ts
@@ -8,3 +8,9 @@ export const KEY_CODES = {
   ARROW_UP: 38,
   ARROW_DOWN: 40,
 };
+
+export const VIDEO_INPUT_QUALITY = {
+  '360p': '360p (nHD) @ 15 fps (600 Kbps max)',
+  '540p': '540p (qHD) @ 15 fps (1.4 Mbps max)',
+  '720p': '720p (HD) @ 15 fps (1.4 Mbps max)',
+};